@@ -1,4 +1,4 @@
-﻿#if NetStandard
+﻿#if NetStandard || NET6_0_OR_GREATER
 namespace PrimS.Telnet.CiTests
 #else
 namespace PrimS.Telnet.Sync.CiTests
