@@ -61,6 +61,7 @@
     /// </summary>
     protected void SendCancel()
     {
+#pragma warning disable CA1031 // Do not catch general exception types
       try
       {
         internalCancellation?.Cancel();
@@ -69,6 +70,7 @@
       {
         System.Diagnostics.Debug.WriteLine(ex.Message);
       }
+#pragma warning restore CA1031 // Do not catch general exception types
     }
 
     /// <summary>
@@ -89,8 +91,10 @@
         internalCancellation.Dispose();
       }
 
-      var are = new AutoResetEvent(false);
-      are.WaitOne(100);
+      using (var are = new AutoResetEvent(false))
+      {
+        are.WaitOne(100);
+      }
     }
   }
 }
